Add support for 2-bay HDD enclosures. The first drive is for WiiU and the second drive is for vWii.
WiiU will recognize only the 1st drive but ignore the 2nd, which works in our favor - WiiU won't even nag you to format the 2nd drive to its proprietary format. usbtoggle is not required. Both drives are adequately powered by an external power brick because all 2-bay enclosures are externally powered. No more Y-cables. (WiiU is very strict about the 500mA USB current limit.)

If you have a RAID enclosure which can be set to RAID 0/1/SPAN/2LUN, set it to 2LUN. On a computer 2LUN simply shows as 2 separate hard drives.

Your homebrew shall remain on an SD card. The WiiU side demands that payload.elf be on an SD card and we can put all vWii homebrews on the same SD card, too. Most homebrews run with IOS 58 which does NOT see the second HDD.

GerbilSoft's code in Nintendont is borrowed to check disk signature. Wii U disks are skipped automatically.

If both disks are readable and contain games, in USB Loader GX, set USB Port to 1 to choose the second hard drive. Until now, the USB Port setting only works with an 'alt' cIOS such as d2x-v10-beta53-alt. The usual advice is to use non-alt cIOS and plug the hard drive to USB port 0. That advice still holds, the only change being that the same USB port 0 is now home to 2 HDDs, with the first hard drive appearing as 'USB Port 0' and the second hard drive appearing as 'USB Port 1'.

Duplicates of binaries are made so that they will show up as different groups in d2x-cios-installer. In my opinion this is the less risky solution in the wrong hands. Leseratte added console= and region= tags to ciosmaps.xml. A new d2x-cios-installer is needed with enhancements to support those new tags and filter the user interface so that only the options pertinent to the user's hardware are shown. However, nothing stops users from using this cIOS with existing the d2x-cios-installer with no such support, and then proceed to install the wrong cIOS on the wrong console (e.g. Wii cIOS on vWii). The safest thing to do IMO is to group them into ciosgroups. Time comes when somebody makes a new d2x-cios-installer it can still support Leseratte's new tags and filter the options. The only cost to this arrangement is file duplication and extra disk/SD card space.

Upped version to 9 to convince WiiXplorer that USB Set Port API is available.

There has been a watchdog thread that reads a random sector every 10 seconds to keep the hard drive awake but it does not have to move the head across the platter that much. Any close-by sector will do.

Change retry to increase timeout exponentially, from 2, 4, to 8 seconds. Minimum is uniformly 2 seconds based on a WD Scorpio Blue 1TB that I pulled from a WD NAS/Router.

The delay from USB reset to get descriptor is increased for broader drive/controller compatibility.

__cycle() calls __usbstorage_reset() to handle timeouts, but __usbstorage_reset() calls __cycle() to do stuff that has the potential to timeout. Infinite reentrancy. Most of the code is exactly the same as before (d2x-v10), keeping only fixes that are demostrably helpful. Namely, some timeouts are increased, and timeout keeps on increasing as retries are attempted, so that we won't run into a vicious cycle where some operation is ramping up but we keep killing it and restarting it because the timeout we give is too stingy. Secondly, the error handling for an interface failure and a single drive timeout are separate cases. The former continues to use the nuclear strategy of unplugging and hard-resetting the USB interface, but that shall not be applied to the latter. We must not kill the innocent along with the guilty. A soft reset is the appropriate way to deal with a single drive timeout.